+++ src/Cli/Console.php
@@ -43,6 +43,8 @@ final class Console
         $cli->addCommand(new PagePreviewCommand($this->app));
         $cli->addCommand(new PageListCommand($this->app));
         $cli->addCommand(new PageShowCommand($this->app));
+        $cli->addCommand(new PageDraftCommand($this->app));
+        $cli->addCommand(new PagePublishCommand($this->app));
         $cli->addCommand(new IdGenerateCommand($this->app));
 
         return $cli->run(null, $output);